Модуль 5. Задание 2. (Правки от наставника)
