+++ src/components/sort.js
@@ -1,14 +1,5 @@
 import AbstractComponent from './abstract-component.js';
-
-const getSortTemplate = () => {
-  return (
-    `<ul class="sort">
-    <li><a href="#" class="sort__button sort__button--active">Sort by default</a></li>
-    <li><a href="#" class="sort__button">Sort by date</a></li>
-    <li><a href="#" class="sort__button">Sort by rating</a></li>
-  </ul>`
-  );
-};
+import {getSortTemplate} from '../templates/sort.js';
 
 export default class Sort extends AbstractComponent {
   getTemplate() {
