+++ src/components/user-rank.js
@@ -1,13 +1,5 @@
 import AbstractComponent from './abstract-component.js';
-
-export const getUserRankTemplate = (rank) => {
-  return (
-    `<section class="header__profile profile">
-    <p class="profile__rating">${rank}</p>
-    <img class="profile__avatar" src="images/bitmap@2x.png" alt="Avatar" width="35" height="35">
-  </section>`
-  );
-};
+import {getUserRankTemplate} from '../templates/user-rank.js';
 
 export default class UserRank extends AbstractComponent {
   constructor(rank = 0) {
