+++ src/main.js
@@ -3,7 +3,6 @@ import UserRankComponent from './components/user-rank.js';
 import {generateCards} from './mock/card.js';
 import {getUserRank} from './mock/user-rank.js';
 import PageController from './controllers/page-controller';
-import {render} from './utils/render.js';
 import {FilmCount, RenderPosition} from './utils/const.js';
 
 const headerElement = document.querySelector(`.header`);
@@ -13,10 +12,12 @@ const mainElement = document.querySelector(`.main`);
 const cards = generateCards(FilmCount.ALL);
 
 // Рендер ранга юзера
-render(headerElement, new UserRankComponent(getUserRank()), RenderPosition.BEFOREEND);
+const userRankComponent = new UserRankComponent(getUserRank());
+userRankComponent.render(headerElement, RenderPosition.BEFOREEND);
 
 // Рендер сортировки
-render(mainElement, new SortComponent(), RenderPosition.BEFOREEND);
+const sortComponent = new SortComponent();
+sortComponent.render(mainElement, RenderPosition.BEFOREEND);
 
 // Установка числа всех фильмов в футер
 footerElement.querySelector(`.footer__statistics p`).textContent = `${cards.length} movies inside`;
